Update BookRegistry and BookRegistryRecord for audiobook bookmark
