+++ Simplified/Book/Models/NYPLBookRegistryRecord.h
@@ -3,6 +3,7 @@
 @class NYPLBook;
 @class NYPLBookLocation;
 @class NYPLReadiumBookmark;
+@class NYPLAudiobookBookmark;
 typedef NS_ENUM(NSInteger, NYPLBookState);
 
 @interface NYPLBookRegistryRecord : NSObject
@@ -12,6 +13,7 @@ typedef NS_ENUM(NSInteger, NYPLBookState);
 @property (nonatomic, readonly) NYPLBookState state;
 @property (nonatomic, readonly) NSString *fulfillmentId; // nilable
 @property (nonatomic, readonly) NSArray<NYPLReadiumBookmark *> *readiumBookmarks; // nilable
+@property (nonatomic, readonly) NSArray<NYPLAudiobookBookmark *> *audiobookBookmarks; // nilable
 @property (nonatomic, readonly) NSArray<NYPLBookLocation *> *genericBookmarks; // nilable
 
 + (id)new NS_UNAVAILABLE;
@@ -23,6 +25,7 @@ typedef NS_ENUM(NSInteger, NYPLBookState);
                        state:(NYPLBookState)state
                fulfillmentId:(NSString *)fulfillmentId
             readiumBookmarks:(NSArray<NYPLReadiumBookmark *> *)readiumBookmarks
+          audiobookBookmarks:(NSArray<NYPLAudiobookBookmark *> *)audiobookBookmarks
             genericBookmarks:(NSArray<NYPLBookLocation *> *)genericBookmarks;
 
 // designated initializer
@@ -39,6 +42,8 @@ typedef NS_ENUM(NSInteger, NYPLBookState);
 - (instancetype)recordWithFulfillmentId:(NSString *)fulfillmentId;
 
 - (instancetype)recordWithReadiumBookmarks:(NSArray<NYPLReadiumBookmark *> *)bookmarks;
+
+- (instancetype)recordWithAudiobookBookmarks:(NSArray<NYPLAudiobookBookmark *> *)bookmarks;
 
 - (instancetype)recordWithGenericBookmarks:(NSArray<NYPLBookLocation *> *)bookmarks;
